Отправка и создание сообщений перенесено в ClientNetSend. idNet сменён с int на long.

+++ src/main/player/enemy/EnemyBullet.java
@@ -5,9 +5,9 @@ import main.*;
 public class EnemyBullet extends Obj{
 	
 	public String enemyName;
-	public int idNet;
+	public long idNet;
 	
-	public EnemyBullet(double x, double y, double speed, double direction,Sprite sprite, String enemyName, int idNet,Game game){
+	public EnemyBullet(double x, double y, double speed, double direction,Sprite sprite, String enemyName, long idNet,Game game){
 		super(x,y,speed,direction,0,true,sprite,game);
 		this.enemyName = enemyName;
 		this.idNet = idNet;
